+++ include/boost/phoenix/core/detail/function_eval.hpp
@@ -13,7 +13,7 @@
 #include <boost/phoenix/core/call.hpp>
 #include <boost/phoenix/core/expression.hpp>
 #include <boost/phoenix/core/meta_grammar.hpp>
-#include <boost/phoenix/core/detail/phx2_result.hpp>
+// #include <boost/phoenix/core/detail/phx2_result.hpp>
 #include <boost/utility/result_of.hpp>
 
 #if defined(BOOST_NO_CXX11_VARIADIC_TEMPLATES) || \
@@ -64,7 +64,7 @@ namespace boost { namespace phoenix {
             template <typename Sig>
             struct result;
 
-#ifndef BOOST_PHOENIX_NO_VARIADIC_FUNCTION_EVAL
+#ifdef BOOST_PHOENIX_NO_VARIADIC_FUNCTION_EVAL
             template <typename This, typename F, typename Context>
             struct result<This(F, Context)>
             {
